Finish bloglist expansion step9

+++ part4/controllers/blogs.js
@@ -3,6 +3,12 @@ const blogsRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
 
+blogsRouter.get("/info", async (request, response) => {
+    const count = await Blog.countDocuments();
+    // const count = await Blog.where({ author: "Jane Doe" }).count();
+    response.status(200).json(`Total blogs: ${count}`);
+});
+
 blogsRouter.get("/", async (request, response) => {
     const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
     response.json(blogs);
@@ -39,8 +45,21 @@ blogsRouter.post("/", async (request, response) => {
 });
 
 blogsRouter.delete("/:id", async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id);
-    response.status(204).end();
+    const decodedToken = jwt.verify(request.token, process.env.SECRET);
+
+    if (!decodedToken.id) {
+        return response.status(401).json({ error: "token invalid" });
+    }
+    const user = await User.findById(decodedToken.id);
+    console.log(user.id);
+
+    const blog = await Blog.findById(request.params.id);
+    if (blog.user.toString() === user.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id);
+        response.status(204).end();
+    } else {
+        return response.status(401).json({ error: "Only the blog creator has rights to delete the blog." });
+    }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
